refactor: migrate to tsx
